Uploaded Game files from Godot
This is every single file inside of my Godot folder including the game and config. The gitignore file is set to ignore .tmp files which is a change from the default file which gets made. .tmp files are temp save files and including them in uploading to github can cause issues with overwriting and conflicting file names.