Addressed Issue #7 and #9
Re-uploaded files in whole (except temp save data).

Created an object in level_0, and a folder dedicated to the player.

Inside of the player folder created player scene and script to handle basic movement, but also to implement the ratone.blend file, re-exported as a .obj file, then put into level_0 as an instanced scene.